moved command buffer out of renderer for cleanlieness
