+++ src/rendering/dm_renderer.c
@@ -1,4 +1,5 @@
 #include "dm_renderer.h"
+#include "dm_command_buffer.h"
 #include "dm_mem.h"
 #include "dm_logger.h"
 
@@ -29,17 +30,6 @@ void dm_renderer_bind_shader_impl(dm_shader* shader);
 
 bool dm_renderer_create_render_pipeline_impl(dm_render_pipeline* pipeline);
 void dm_renderer_destroy_render_pipeline_impl(dm_render_pipeline* pipeline);
-
-void dm_renderer_begin_renderpass_impl(dm_render_pipeline* pipeline);
-void dm_renderer_end_rederpass_impl();
-bool dm_renderer_bind_pipeline_impl(dm_render_pipeline* pipeline);
-void dm_renderer_bind_vertex_buffer_impl(dm_buffer* buffer);
-void dm_renderer_bind_index_buffer_impl(dm_buffer* buffer);
-void dm_renderer_set_viewport_impl(dm_viewport* viewport);
-void dm_renderer_clear_impl(dm_color* clear_color);
-
-void dm_renderer_draw_arrays_impl(int first, size_t count);
-void dm_renderer_draw_indexed_impl(dm_draw_indexed_params* params, dm_render_pipeline* pipeline);
 
 // renderer resources; buffers, shaders, etc
 static dm_render_resources resources;
@@ -131,7 +121,7 @@ bool dm_renderer_end_scene()
 {
 	dm_renderer_submit_command(DM_RENDER_COMMAND_END_RENDER_PASS, NULL, &r_data.object_pipeline->command_buffer);
 
-	if (!dm_renderer_submit_command_buffer(&r_data.object_pipeline->command_buffer)) return false;
+	if (!dm_renderer_submit_command_buffer(&r_data.object_pipeline->command_buffer, r_data.object_pipeline)) return false;
 
 	dm_renderer_end_scene_impl(&r_data);
 
@@ -244,8 +234,8 @@ bool dm_renderer_init_object_data()
 	};
 
 	// buffers
-	dm_buffer_desc v_desc = { .type = DM_BUFFER_TYPE_VERTEX, .size = sizeof(vertices) };
-	dm_buffer_desc i_desc = { .type = DM_BUFFER_TYPE_INDEX, .size = sizeof(indices) };
+	dm_buffer_desc v_desc = { .type = DM_BUFFER_TYPE_VERTEX, .size = sizeof(vertices), .usage=DM_BUFFER_USAGE_STATIC };
+	dm_buffer_desc i_desc = { .type = DM_BUFFER_TYPE_INDEX, .size = sizeof(indices), .usage=DM_BUFFER_USAGE_STATIC };
 
 	if (!dm_renderer_create_buffer(v_desc, &r_data.object_pipeline->render_packet.vertex_buffer)) return false;
 	if (!dm_renderer_create_buffer(i_desc, &r_data.object_pipeline->render_packet.index_buffer)) return false;
@@ -381,61 +371,4 @@ dm_shader* dm_renderer_get_shader(dm_shader_handle handle)
 		DM_LOG_ERROR("Trying to access invalid shader!");
 		return NULL;
 	}
-}
-
-void dm_renderer_submit_command(dm_render_command_type command_type, void* data, dm_command_buffer* command_buffer)
-{
-	dm_render_command command = { .command = command_type, .data = data };
-	dm_list_append(&command_buffer->commands, command);
-}
-
-void dm_renderer_clear_command_buffer(dm_command_buffer* command_buffer)
-{
-	dm_list_clear(&command_buffer->commands);
-}
-
-void dm_renderer_destroy_command_buffer(dm_command_buffer* command_buffer)
-{
-	dm_list_destroy(&command_buffer->commands);
-}
-
-bool dm_renderer_submit_command_buffer(dm_command_buffer* command_buffer)
-{
-	dm_list_for_range(command_buffer->commands)
-	{
-		dm_render_command command = command_buffer->commands.array[i];
-
-		switch (command.command)
-		{
-		// TODO flesh out
-		case DM_RENDER_COMMAND_BEGIN_RENDER_PASS:
-		{
-			dm_renderer_begin_renderpass_impl(r_data.object_pipeline);
-		} break;
-		case DM_RENDER_COMMAND_END_RENDER_PASS:
-		{
-			dm_renderer_end_rederpass_impl();
-		} break;
-		case DM_RENDER_COMMAND_SET_VIEWPORT:
-		{
-			dm_renderer_set_viewport_impl(&r_data.object_pipeline->viewport);
-		} break;
-		case DM_RENDER_COMMAND_CLEAR:
-		{
-			dm_renderer_clear_impl((dm_color*)command.data);
-		} break;
-		case DM_RENDER_COMMAND_BIND_PIPELINE:
-		{
-			if (!dm_renderer_bind_pipeline_impl((dm_render_pipeline*)command.data)) return false;
-		} break;
-		case DM_RENDER_COMMAND_DRAW_INDEXED:
-		{
-			dm_renderer_draw_indexed_impl((dm_draw_indexed_params*)command.data, r_data.object_pipeline);
-		} break;
-		case DM_RENDER_COMMAND_DRAW_INSTANCED:
-		{} break;
-		}
-	}
-
-	return true;
 }
